delete category fake data file and merged it with one dummy data object
