+++ app/src/test/java/com/baghdad/tudee/data/service/CategoryServiceImplTest.kt
@@ -2,19 +2,19 @@ package com.baghdad.tudee.data.service
 
 import android.database.sqlite.SQLiteException
 import com.baghdad.tudee.data.database.dao.CategoryDao
-import com.baghdad.tudee.data.service.SampleCategoryData.categoryID
-import com.baghdad.tudee.data.service.SampleCategoryData.dbError
-import com.baghdad.tudee.data.service.SampleCategoryData.duplicateCategoryTitle
-import com.baghdad.tudee.data.service.SampleCategoryData.emptyCategoryList
-import com.baghdad.tudee.data.service.SampleCategoryData.expectedImageUri
-import com.baghdad.tudee.data.service.SampleCategoryData.expectedTitle
-import com.baghdad.tudee.data.service.SampleCategoryData.index
-import com.baghdad.tudee.data.service.SampleCategoryData.invalidCategoryData
-import com.baghdad.tudee.data.service.SampleCategoryData.invalidCategoryId
-import com.baghdad.tudee.data.service.SampleCategoryData.nonExistentCategoryId
-import com.baghdad.tudee.data.service.SampleCategoryData.oneCategoryExpected
-import com.baghdad.tudee.data.service.SampleCategoryData.sampleCategory
-import com.baghdad.tudee.data.service.SampleCategoryData.sampleDto
+import com.baghdad.tudee.data.service.SampleTaskData.categoryID
+import com.baghdad.tudee.data.service.SampleTaskData.dbError
+import com.baghdad.tudee.data.service.SampleTaskData.duplicateCategoryTitle
+import com.baghdad.tudee.data.service.SampleTaskData.emptyCategoryList
+import com.baghdad.tudee.data.service.SampleTaskData.expectedImageUri
+import com.baghdad.tudee.data.service.SampleTaskData.expectedTitle
+import com.baghdad.tudee.data.service.SampleTaskData.index
+import com.baghdad.tudee.data.service.SampleTaskData.invalidCategoryData
+import com.baghdad.tudee.data.service.SampleTaskData.invalidCategoryId
+import com.baghdad.tudee.data.service.SampleTaskData.nonExistentCategoryId
+import com.baghdad.tudee.data.service.SampleTaskData.oneCategoryExpected
+import com.baghdad.tudee.data.service.SampleTaskData.sampleCategory
+import com.baghdad.tudee.data.service.SampleTaskData.sampleDto
 import com.baghdad.tudee.domain.exception.DatabaseException
 import io.mockk.coEvery
 import io.mockk.coVerify
